+++ tests/Unit/DSC_WebConfigPropertyCollection.Tests.ps1
@@ -245,6 +245,7 @@ try
             Context 'Ensure is present and collection item and property do not exist - String ItemPropertyValue' {
                 Mock -CommandName Get-ItemValues -ModuleName $script:dscResourceName
                 Mock -CommandName Add-WebConfigurationProperty
+                Mock -CommandName Set-WebConfigurationProperty
                 Mock -CommandName Get-CollectionItemPropertyType -MockWith { return 'String' }
                 Mock -CommandName Convert-PropertyValue
 
@@ -253,6 +254,7 @@ try
 
                     Assert-MockCalled -CommandName Get-ItemValues -Times 1 -Exactly
                     Assert-MockCalled -CommandName Add-WebConfigurationProperty -Times 1 -Exactly
+                    Assert-MockCalled -CommandName Set-WebConfigurationProperty -Times 0 -Exactly
                     Assert-MockCalled -CommandName Get-CollectionItemPropertyType -Times 1 -Exactly
                     Assert-MockCalled -CommandName Convert-PropertyValue -Times 0 -Exactly
                 }
@@ -261,6 +263,7 @@ try
             Context 'Ensure is present and collection item and property do not exist - Integer ItemPropertyValue' {
                 Mock -CommandName Get-ItemValues -ModuleName $script:dscResourceName
                 Mock -CommandName Add-WebConfigurationProperty
+                Mock -CommandName Set-WebConfigurationProperty
                 Mock -CommandName Get-CollectionItemPropertyType -MockWith { return 'Int64' }
                 Mock -CommandName Convert-PropertyValue
 
@@ -269,6 +272,7 @@ try
 
                     Assert-MockCalled -CommandName Get-ItemValues -Times 1 -Exactly
                     Assert-MockCalled -CommandName Add-WebConfigurationProperty -Times 1 -Exactly
+                    Assert-MockCalled -CommandName Set-WebConfigurationProperty -Times 0 -Exactly
                     Assert-MockCalled -CommandName Get-CollectionItemPropertyType -Times 1 -Exactly
                     Assert-MockCalled -CommandName Convert-PropertyValue -Times 1 -Exactly
                 }
@@ -281,6 +285,7 @@ try
                         allowed = 'false'
                     }
                 }
+                Mock -CommandName Add-WebConfigurationProperty
                 Mock -CommandName Set-WebConfigurationProperty -MockWith {}
                 Mock -CommandName Get-CollectionItemPropertyType -MockWith { return 'String' }
                 Mock -CommandName Convert-PropertyValue
@@ -289,6 +294,7 @@ try
                     Set-TargetResource @script:presentParameters
 
                     Assert-MockCalled -CommandName Get-ItemValues -Times 1 -Exactly
+                    Assert-MockCalled -CommandName Add-WebConfigurationProperty -Times 0 -Exactly
                     Assert-MockCalled -CommandName Set-WebConfigurationProperty -Times 1 -Exactly
                     Assert-MockCalled -CommandName Get-CollectionItemPropertyType -Times 1 -Exactly
                     Assert-MockCalled -CommandName Convert-PropertyValue -Times 0 -Exactly