+++ CHANGELOG.md
@@ -7,6 +7,15 @@ For older change log history see the [historic changelog](HISTORIC_CHANGELOG.md)
 
 ## [Unreleased]
 
+### Added
+
+- WebConfigPropertyCollection
+  - Allowed different property collection key types to be added beyond the default.
+- IisModule
+  - Set default for Ensure property to Present.
+- IisMimeTypeMapping
+  - Set default for Ensure property to Present.
+
 ## [4.2.0] - 2024-08-26
 
 ### Removed