+++ CHANGELOG.md
@@ -11,6 +11,9 @@ For older change log history see the [historic changelog](HISTORIC_CHANGELOG.md)
 
 - WebConfigPropertyCollection
   - Allowed different property collection key types to be added beyond the default.
+
+### Changed
+
 - IisModule
   - Set default for Ensure property to Present.
 - IisMimeTypeMapping